+++ src/config.rs
@@ -6,15 +6,15 @@ use std::error::Error;
 use std::fmt;
 
 fn default_db_path() -> String {
-    "/var/renews/news.db".into()
+    "sqlite:///var/renews/news.db".into()
 }
 
 fn default_auth_db_path() -> String {
-    "/var/renews/auth.db".into()
+    "sqlite:///var/renews/auth.db".into()
 }
 
 fn default_peer_db_path() -> String {
-    "/var/renews/peers.db".into()
+    "sqlite:///var/renews/peers.db".into()
 }
 
 fn default_peer_sync_secs() -> u64 {
